Playlist Editor: Rich text destroyed when dragging relates to xibosignageltd/xibo-private#431

+++ ui/src/playlist-editor/main.js
@@ -717,8 +717,9 @@ pE.showLocalLoadingScreen = function() {
 
 /**
  * Clear Temporary Data ( Cleaning cached variables )
+ * @param {Boolean} destroyRichTextEditor Destroy existing Rich text editors
  */
-pE.clearTemporaryData = function() {
+pE.clearTemporaryData = function(destroyRichTextEditor = false) {
   // Fix for remaining ckeditor elements or colorpickers
   destroyColorPicker(pE.editorContainer.find('.colorpicker-element'));
 
@@ -726,7 +727,7 @@ pE.clearTemporaryData = function() {
   pE.editorContainer.find('.tooltip').remove();
 
   // Remove text callback editor structure variables
-  formHelpers.destroyCKEditor();
+  (destroyRichTextEditor) && formHelpers.destroyCKEditor();
 };
 
 /**
